Graph is no longer encoded in the filesystem
Encoding the path taken in the URL in a graph is basically nonsense,
if you allow for traversing **up** as well as down the graph. You can
move to a parent-node that is entirely seperate from the path-history
you to took, except that child node joining them.

The only solution is to pick some random path to reach the parent node,
and rewrite the URL to it... which kills the point of having the path recorded.

If we're removing the URL-recording, then there's also no reason to have
the symlink system either. So that's also gone out the window. With the
removal of the symlinks, a large chunk of the post_graph.rs code has
become dead code, and should be removed (but hasn't been yet).

This also means that we can move the output up to the root dir, and get
rid of src/.

Also added support for directories in the output directory, with new
field `dirname` in the post metadata.

So finally, the graph is now encoded by the HTML links, and not in the
filesystem. This also means the JS code is unecessary, and was removed.

:-)